"Ecommerce project using SpringBoot with swagger"

+++ src/main/java/com/yudhi/ecom_project/controller/ProductController.java
@@ -28,7 +28,7 @@ public class ProductController {
 
     }
     @GetMapping("/product/{id}")
-    public ResponseEntity<Product> getProductById(@PathVariable long id){
+    public ResponseEntity<Product> getProductById(@PathVariable int id){
         Product product = service.getProductById(id);
         if(product != null)
             return new ResponseEntity<>(product, HttpStatus.OK);
@@ -48,7 +48,7 @@ public class ProductController {
 
         }
     @GetMapping("product/{productid}/image")
-    public ResponseEntity<byte[]> getImageById(@PathVariable long productid){
+    public ResponseEntity<byte[]> getImageById(@PathVariable int productid){
         Product product = service.getProductById(productid);
         byte[]imageFile=product.getImageDate();
         return ResponseEntity.ok()
@@ -57,7 +57,7 @@ public class ProductController {
     }
 
     @PutMapping("/product/{id}")
-    public ResponseEntity<String>updateProduct(@PathVariable long id,@RequestPart Product product,
+    public ResponseEntity<String>updateProduct(@PathVariable int  id,@RequestPart Product product,
                                                @RequestPart MultipartFile imageFile){
         Product product1 = null;
         try {
@@ -73,7 +73,7 @@ public class ProductController {
 
     //delete
     @DeleteMapping("/product/{id}")
-    public ResponseEntity<String>deleteProduct(@PathVariable long id){
+    public ResponseEntity<String>deleteProduct(@PathVariable int id){
         Product product = service.getProductById(id);
         if(product != null){
             service.deleteProduct(id);
